+++ packages/camera/camera_android_camerax/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.5.0+18
+
+* Implements `startVideoCapturing`.
+
+## 0.5.0+17
+
+* Implements resolution configuration for all camera use cases.
+
 ## 0.5.0+16
 
 * Adds pub topics to package metadata.